requirements.txt
SpeechRecognition
pyttsx3
python-dotenv
pyaudio
setuptools
google-generativeai
ipython
google-colab
